Move ColVals and impl to iml-influx crate.
