+++ iml-influx/Cargo.toml
@@ -8,3 +8,4 @@ edition = "2018"
 [dependencies]
 iml-wire-types = { path = "../iml-wire-types", version = "0.3" }
 serde = { version = "1", features = ['derive'] }
+serde_json = { version = "1" }
