+++ iml-influx/src/lib.rs
@@ -7,6 +7,8 @@ use std::collections::HashMap;
 pub mod filesystem;
 pub mod filesystems;
 
+use serde_json::{Map, Value};
+
 #[derive(serde::Deserialize, Clone, Debug)]
 pub struct InfluxResponse<T> {
     results: Vec<InfluxResult<T>>,
@@ -21,4 +23,20 @@ pub struct InfluxResult<T> {
 pub struct InfluxSeries<T> {
     tags: Option<HashMap<String, String>>,
     values: Vec<T>,
+}
+
+pub struct ColVals(pub Vec<String>, pub Vec<Vec<serde_json::Value>>);
+
+impl From<ColVals> for serde_json::Value {
+    fn from(ColVals(cols, vals): ColVals) -> Self {
+        let xs = vals
+            .into_iter()
+            .map(|y| -> Map<String, serde_json::Value> {
+                cols.clone().into_iter().zip(y).collect()
+            })
+            .map(Value::Object)
+            .collect();
+
+        Value::Array(xs)
+    }
 }
